BLOG APP

+++ blog-app-apis/blog-app-apis/src/main/java/com/nishant/blog_app_apis/exceptions/ResourceNotFoundException.java
@@ -6,7 +6,7 @@ public class ResourceNotFoundException extends RuntimeException{
     String fieldName;
     long fieldValue;
 
-    public ResourceNotFoundException( String resourceName ,String fieldName, long fieldValue) {
+    public ResourceNotFoundException( String resourceName ,String fieldName, int fieldValue) {
         super(String.format("%s not found with this name %s : %d",resourceName,fieldName,fieldValue));
         this.fieldValue = fieldValue;
         this.fieldName = fieldName;
